create_generated_dataset seems to work, only missing molecular refactor dataseting utils to dataseting_utils.py
