+++ learning_lidar/dataseting/dataseting.py
@@ -1,26 +1,22 @@
 # %% #General Imports
 import os
 from datetime import datetime, timedelta  # time,
-import multiprocess as mp
 from IPython.display import display
 import logging
 from tqdm import tqdm
 
 # %% Scientific and data imports
-import sqlite3
 import numpy as np
 import pandas as pd
 import xarray as xr
-from scipy.ndimage import gaussian_filter1d
-from scipy.interpolate import griddata
-from functools import partial
 
 # %% Local modules imports
 import learning_lidar.utils.global_settings as gs
 import learning_lidar.preprocessing.preprocessing as prep
+from learning_lidar.dataseting.dataseting_utils import get_query, query_database, add_profiles_values, add_X_path, \
+    get_time_slots_expanded, convert_Y_features_units, recalc_LC, split_save_train_test_ds, get_generated_X_path, \
+    get_mean_lc
 from learning_lidar.utils.utils import create_and_configer_logger
-
-eps = np.finfo(np.float).eps
 
 
 # %% Dataset creating helper functions
@@ -117,139 +113,6 @@ def create_dataset(station_name='haifa', start_date=datetime(2017, 9, 1),
     return full_df.reset_index(drop=True)
 
 
-def get_query(wavelength, cali_method, day_date):
-    start_time = datetime.combine(date=day_date.date(), time=day_date.time().min)
-    end_time = datetime.combine(date=day_date.date(), time=day_date.time().max)
-    query = f"""
-    SELECT  lcc.liconst, lcc.uncertainty_liconst,
-            lcc.cali_start_time, lcc.cali_stop_time,
-            lcc.wavelength, lcc.cali_method, lcc.telescope
-    FROM lidar_calibration_constant as lcc
-    WHERE
-        wavelength == {wavelength} AND
-        cali_method == '{cali_method}' AND
-        telescope == 'far_range' AND
-        (cali_start_time BETWEEN '{start_time}' AND '{end_time}');
-    """
-    return query
-
-
-def query_database(query="SELECT * FROM lidar_calibration_constant;",
-                   database_path="pollyxt_tropos_calibration.db"):
-    """
-    Query is a string following sqlite syntax (https://www.sqlitetutorial.net/) to query the .db
-    Examples:
-    query_basic = "
-    SELECT * -- This is a comment. Get all columns from table
-    FROM lidar_calibration_constant -- Which table to query
-    "
-
-    query_advanced = "
-    SELECT lcc.id, lcc.liconst, lcc.cali_start_time, lcc.cali_stop_time -- get only some columns
-    FROM lidar_calibration_constant as lcc
-    WHERE -- different filtering options on rows
-        wavelength == 1064 AND
-        cali_method LIKE 'Klet%' AND
-        (cali_start_time BETWEEN '2017-09-01' AND '2017-09-02');
-    "
-    """
-    # logger = logging.getLogger ( )
-
-    # Connect to the db and query it directly into pandas df.
-    # try:
-    with sqlite3.connect(database_path) as c:
-        # Query to df
-        # optionally parse 'id' as index column and 'cali_start_time', 'cali_stop_time' as dates
-        df = pd.read_sql(sql=query, con=c, parse_dates=['cali_start_time', 'cali_stop_time'])
-    # except sqlite3.OperationalError as e:
-    #    logger.exception ( f"{e}: Unable to load dataset." )
-    #    sys.exit(1)
-    # TODO: raise load exception if file does not exsits. the above commented solution is not really catching this and continues to run
-
-    return df
-
-
-def add_profiles_values(df, station, day_date, file_type='profiles'):
-    logger = logging.getLogger()
-    try:
-        df['matched_nc_profile'] = df.apply(lambda row: prep.get_TROPOS_dataset_paths(station, day_date,
-                                                                                      start_time=row.cali_start_time,
-                                                                                      end_time=row.cali_stop_time,
-                                                                                      file_type=file_type)[
-            0],
-                                            axis=1, result_type='expand')
-    except Exception:
-        logger.exception(
-            f"Non resolved 'matched_nc_profile' for {station.location} station, at date {day_date.strftime('%Y-%m-%d')} ")
-        pass
-
-    def _get_info_from_profile_nc(row):
-        """
-        Get the r_0,r_1, and delta_r of the selected row. The values are following rebasing according to sea-level height.
-        :param row:
-        :return:
-        """
-        data = prep.load_dataset(row['matched_nc_profile'])
-        wavelen = row.wavelength
-        # get altitude to rebase the reference heights according to sea-level-height
-        altitude = data.altitude.item()
-        [r0, r1] = data[f'reference_height_{wavelen}'].values
-        [bin_r0, bin_r1] = [np.argmin(abs(data.height.values - r)) for r in [r0, r1]]
-        delta_r = r1 - r0
-        return r0 + altitude, r1 + altitude, delta_r, bin_r0, bin_r1
-
-    df[['r0', 'r1', 'dr', 'bin_r0', 'bin_r1']] = df.apply(_get_info_from_profile_nc, axis=1,
-                                                          result_type='expand')
-    return df
-
-
-def add_X_path(df, station, day_date, lambda_nm=532, data_source='molecular', file_type='attbsc'):
-    """
-    Add path of 'molecular' or 'lidar' dataset to df
-    :param df: pd.DataFrame ( ) result from database query
-    :param station: gs.station() object of the lidar station
-    :param day_date: datetime.datetime object of the measuring date
-    :param lambda_nm: wavelength [nm] e.g., for the green channel 532 [nm]
-    :param data_source: string object: 'molecular' or 'lidar'
-    :param file_type: string object: e.g., 'attbsc' for molecular_dataset or 'range_corr' for a lidar_dataset
-    :return: ds with the added collum of the relevant raw
-    """
-
-    paths = prep.get_prep_dataset_paths(station=station,
-                                        day_date=day_date,
-                                        data_source=data_source,
-                                        lambda_nm=lambda_nm,
-                                        file_type=file_type)
-    if not paths:
-        raise Exception(
-            f"\n Not exiting any '{data_source}' path for {station.location} station, at {day_date.strftime('%Y-%m-%d')}")
-    elif len(paths) != 1:
-        raise Exception(
-            f"\n Expected ONE '{data_source}' path for {station.location} station, at {day_date.strftime('%Y-%m-%d')}.\nGot: {paths} ")
-    df[f"{data_source}_path"] = paths[0]
-    return df
-
-
-def get_time_slots_expanded(df, sample_size):
-    if sample_size:
-        expanded_df = pd.DataFrame()
-        for indx, row in df.iterrows():
-            time_slots = pd.date_range(row.loc['cali_start_time'], row.loc['cali_stop_time'], freq=sample_size)
-            time_slots.freq = None
-            if len(time_slots) < 2:
-                continue
-            for start_time, end_time in zip(time_slots[:-1], time_slots[1:]):
-                mini_df = row.copy()
-                mini_df['start_time_period'] = start_time
-                mini_df['end_time_period'] = end_time
-                expanded_df = expanded_df.append(mini_df)
-    else:
-        expanded_df = df.copy()
-        expanded_df['start_time_period'] = expanded_df['cali_start_time']
-        expanded_df['end_time_period'] = expanded_df['cali_stop_time']
-    return expanded_df
-
-
 # %% Dataset extending helper functions
 def extend_calibration_info(df):
     """
@@ -269,14 +132,6 @@ def extend_calibration_info(df):
     df['rm'] = df[['r0', 'r1']].progress_apply(np.mean, axis=1, result_type='expand').astype(
         float)
     logger.info(f"Done database extension")
-    return df
-
-
-def convert_Y_features_units(df, Y_features=['LC', 'LC_std', 'r0', 'r1', 'dr'],
-                             scales={'LC': 1E-9, 'LC_std': 1E-9, 'r0': 1E-3, 'r1': 1E-3, 'dr': 1E-3}):
-    Y_scales = [scales[feature] for feature in Y_features]
-    for feature, scale in zip(Y_features, Y_scales):
-        df[feature] *= scale
     return df
 
 
@@ -336,138 +191,54 @@ def create_calibration_ds(df, station, source_file):
     return ds_calibration
 
 
-def recalc_LC(row):
+def create_generated_dataset(station, start_date, end_date, sample_size='30min'):
     """
-	Extending database:by recalculate LC_recalc, LCp_recalc
-	:param row:
-	:return:
-	"""
-    sliced_ds, _ = get_sample_ds(row)
-    slice_ratio = sliced_ds[0].copy(deep=True)
-    slice_ratio.data /= sliced_ds[1].data
-    LC_recalc = slice_ratio.mean().values.item()
-    LCp_recalc = slice_ratio.where(slice_ratio >= 0.0).mean().values.item()
-    return LC_recalc, LCp_recalc
-
-
-# %% General datasets functions
-def _df_split2proc(tup_arg, **kwargs):
-    split_ind, df_split, df_f_name = tup_arg
-    return (split_ind, getattr(df_split, df_f_name)(**kwargs))
-
-
-def df_multi_core(df, df_f_name, subset=None, njobs=-1, **kwargs):
-    # %% testing multiproccesing from: https://gist.github.com/morkrispil/3944242494e08de4643fd42a76cb37ee
-    if njobs == -1:
-        njobs = mp.cpu_count()
-    pool = mp.Pool(processes=njobs)
-
-    try:
-        df_sub = df[subset] if subset else df
-        splits = np.array_split(df_sub, njobs)
-    except ValueError:
-        splits = np.array_split(df, njobs)
-
-    pool_data = [(split_ind, df_split, df_f_name) for split_ind, df_split in enumerate(splits)]
-    results = pool.map(partial(_df_split2proc, **kwargs), pool_data)
-    pool.close()
-    pool.join()
-    results = sorted(results, key=lambda x: x[0])
-    results = pd.concat([split[1] for split in results])
-    return results
-
-
-def get_sample_ds(row):
-    mol_path = row.molecular_path
-    lidar_path = row.lidar_path
-    bin_r0 = row.bin_r0
-    bin_r1 = row.bin_r1
-    t0 = row.start_time_period
-    t1 = row.end_time_period
-    full_ds = [prep.load_dataset(path) for path in [lidar_path, mol_path]]
-    tslice = slice(t0, t1)
-    profiles = ['range_corr', 'attbsc']
-    sliced_ds = [ds_i.sel(Time=tslice,
-                          Height=slice(float(ds_i.Height[bin_r0].values),
-                                       float(ds_i.Height[bin_r1].values)))[profile]
-                 for ds_i, profile in zip(full_ds, profiles)]
-    return sliced_ds, full_ds
-
-
-def get_aerBsc_profile_ds(path, profile_df):
-    cur_profile = prep.load_dataset(path)
-    height_units = 'km'
-    height_scale = 1e-3  # converting [m] to [km]
-    bsc_scale = 1e+3  # converting [1/m sr] to to [1/km sr]
-    heights_indx = height_scale * (cur_profile.height.values + cur_profile.altitude.values)
-    start_time = datetime.utcfromtimestamp(cur_profile.start_time.values[0].tolist())
-    end_time = datetime.utcfromtimestamp(cur_profile.end_time.values[0].tolist())
-    time_indx = pd.date_range(start=start_time, end=end_time, freq='30S')
-    profile_r0 = profile_df.groupby(['wavelength'])['bin_r0'].unique().astype(np.int).values
-    profile_r1 = profile_df.groupby(['wavelength'])['bin_r0'].unique().astype(np.int).values
-    ds_chans = []
-
-    var_names = [f"aerBsc_klett_{wavelength}" for wavelength in [355, 532, 1064]]
-    for v_name, wavelength, r in zip(var_names, [355, 532, 1064], profile_r1):
-        vals = cur_profile[v_name].values
-        vals[r:] = eps
-        vals = gaussian_filter1d(vals, 21, mode='nearest')
-        vals = vals.T.reshape(len(heights_indx), 1)
-        vals[vals < 0] = eps
-        vals *= bsc_scale
-        mat_vals = np.repeat(vals, len(time_indx), axis=1)
-        aerbsc_df = pd.DataFrame(mat_vals, index=heights_indx, columns=time_indx)
-        aerBsc_ds_chan = xr.Dataset(
-            data_vars={'aerBsc': (('Height', 'Time'), aerbsc_df),
-                       'lambda_nm': ('Wavelength', [wavelength])
-                       },
-            coords={'Height': aerbsc_df.index.to_list(),
-                    'Time': aerbsc_df.columns,
-                    'Wavelength': [wavelength]
-                    })
-        aerBsc_ds_chan.aerBsc.attrs = {'long_name': r'$\beta$',  # _{{a}}$' ,
-                                       'units': r'$km^{{-1}} sr^{-1}$',
-                                       'info': r'$Aerosol backscatter$'}
-        # set attributes of coordinates
-        aerBsc_ds_chan.Height.attrs = {'units': '{}'.format('{}'.format(height_units)),
-                                       'info': 'Measurements heights above sea level'}
-        aerBsc_ds_chan.Wavelength.attrs = {'long_name': r'$\lambda$', 'units': r'$nm$'}
-        ds_chans.append(aerBsc_ds_chan)
-    return xr.concat(ds_chans, dim='Wavelength')
-
-
-def split_save_train_test_ds(csv_path='', train_size=0.8, df=None):
+    Creates a dataframe consisting of:
+        date | wavelength | start_time | end_time | lidar_path | bg_path | molecular_path | LC
+    :param station:
+    :param start_date:
+    :param end_date:
+    :param sample_size:
+    :return:
     """
-    Splitting a dataset into 2 : train and test sets. Then saving with similar abbreviation of the original dataset.
-    :param train_size:  np.float(). in [0,1] - the ratio of train size from the dataset
-    :param df: pd.DataFrame(). Optional. Dataset of samples for learning calibration method.
-    :return: train_set, test_set. Of type pd.DataFrame().
-    """
-
-    logger = logging.getLogger()
-    if df is None:
+    dates = pd.date_range(start=start_date, end=end_date, freq=sample_size,
+                          closed='left')  # excludes the last date/record
+    wavelengths = gs.LAMBDA_nm().get_elastic()
+    full_df = pd.DataFrame()
+    for wavelength in tqdm(wavelengths):
         try:
-            df = pd.read_csv(csv_path)
+            df = pd.DataFrame()
+            df['date'] = dates.strftime('%Y-%m-%d')  # new row for each time
+            df['wavelength'] = wavelength  # broadcast single value to all rows
+            # start_time is simply the time and end_time is the start_time + 29 minutes and 59 seconds
+            df['start_time'] = dates
+            df['end_time'] = dates + timedelta(minutes=29, seconds=59)
+
+            # add bg and lidar paths
+            df['bg_path'] = df.apply(
+                lambda row: get_generated_X_path(station=station, parent_folder=station.gen_bg_dataset,
+                                                 cur_day=row['start_time'], data_source='bg',
+                                                 wavelength=wavelength, file_type='p_bg'),
+                axis=1, result_type='expand')
+
+            df['lidar_path'] = df.apply(
+                lambda row: get_generated_X_path(station=station, parent_folder=station.gen_lidar_dataset,
+                                                 cur_day=row['start_time'], data_source='lidar',
+                                                 wavelength=wavelength, file_type='range_corr'),
+                axis=1, result_type='expand')
+
+            # TODO molecular path?
+
+            # get the mean LC from signal_paths, one day at a time
+            for cur_day in pd.date_range(start=start_date, end=end_date, freq='D', closed='left'):
+                df = get_mean_lc(df=df, station=station, cur_day=cur_day)
+
+            full_df = full_df.append(df)
         except FileNotFoundError as e:
-            logger.error(f'\n{e}.\nPlease provide a dataset (df) or an exiting csv path (csv_path).')
-
-    source_path = csv_path.split('.csv')[0]
-    train_path = f'{source_path}_train.csv'
-    test_path = f'{source_path}_test.csv'
-
-    df_copy = df.copy(deep=True)
-    train_set = (df_copy.sample(frac=train_size, random_state=2021)).sort_index()
-    test_set = df_copy.drop(train_set.index)
-    train_set['idx'] = train_set.index.values
-    test_set['idx'] = test_set.index.values
-
-    test_set.to_csv(test_path, index=False)
-    train_set.to_csv(train_path, index=False)
-    logger.info(f"\nThe dataset split to train and test datasets, saved to :\n{test_path},{train_path}")
-    return train_set, test_set
+            print(e)
+    return full_df
 
 
-# %% MAIN
 def main(station_name, start_date, end_date):
     logging.getLogger('matplotlib').setLevel(logging.ERROR)  # Fix annoying matplotlib logs
     logging.getLogger('PIL').setLevel(logging.ERROR)  # Fix annoying PIL logs
@@ -480,7 +251,6 @@ def main(station_name, start_date, end_date):
     USE_KM_UNITS = False
     SPLIT_DATASET = False
     DO_GENERATED_DATASET = True
-
 
     # Load data of station
     station = gs.Station(stations_csv_path='../../data/stations.csv', station_name=station_name)
@@ -539,71 +309,9 @@ def main(station_name, start_date, end_date):
         prep.save_dataset(ds_calibration, os.path.curdir, ds_path_extended)
         logger.info(f"The calibration dataset saved to :{os.path.join(os.path.curdir, ds_path_extended)}")
 
-
     if DO_GENERATED_DATASET:
-        create_generated_dataset(station, start_date, end_date)
+        generated_df = create_generated_dataset(station, start_date, end_date)
 
-def create_generated_dataset(station, start_date, end_date, sample_size='30min'):
-    """
-    Creates a dataframe consisting of:
-        date | wavelength | start_time | end_time | lidar_path | bg_path | molecular_path | LC
-    :param station:
-    :param start_date:
-    :param end_date:
-    :param sample_size:
-    :return:
-    """
-    # station.gen_bg_dataset - bg signals
-    # station.gen_lidar_dataset - lidar
-    # station.molecular_dataset - moecular
-    dates = pd.date_range(start=start_date, end=end_date, freq=sample_size)
-    wavelengths = gs.LAMBDA_nm().get_elastic()
-    # TODO
-    """ 
-    for each wavelength, and for each day:
-        row for each start time in day
-    
-    """
-
-    full_df = pd.DataFrame()
-    for wavelength in tqdm(wavelengths):
-        df = pd.DataFrame()
-        df['date'] = dates.strftime('%Y-%m-%d')
-        df['wavelength'] = wavelength
-        df['start_time'] = dates
-        df['end_time'] = dates + timedelta(minutes=29, seconds=59)
-
-        # for day_date in dates:
-        #     try:
-        #         query = get_query(wavelength, cali_method, day_date)
-        #         df = query_database(query=query, database_path=db_path)
-        #         if df.empty:
-        #             raise Exception(
-        #                 f"\n Not existing data for {station.location} station, during {day_date.strftime('%Y-%m-%d')} in '{db_path}'")
-        #         df['date'] = day_date.strftime('%Y-%m-%d')
-        #
-        #         df = add_profiles_values(df, station, day_date, file_type='profiles')
-        #
-        #         df = add_X_path(df, station, day_date, lambda_nm=wavelength, data_source='lidar',
-        #                         file_type='range_corr')
-        #         df = add_X_path(df, station, day_date, lambda_nm=wavelength, data_source='molecular',
-        #                         file_type='attbsc')
-        #
-        #         df = df.rename(
-        #             {'liconst': 'LC', 'uncertainty_liconst': 'LC_std', 'matched_nc_profile': 'profile_path'},
-        #             axis='columns')
-        #         expanded_df = get_time_slots_expanded(df, sample_size)
-        #
-        #         # reorder the columns
-        #         key = ['date', 'wavelength', 'cali_method', 'telescope', 'cali_start_time', 'cali_stop_time',
-        #                'start_time_period', 'end_time_period', 'profile_path']
-        #         Y_features = ['LC', 'LC_std', 'r0', 'r1', 'dr', 'bin_r0', 'bin_r1']
-        #         X_features = ['lidar_path', 'molecular_path']
-        #         expanded_df = expanded_df[key + X_features + Y_features]
-        #         full_df = full_df.append(expanded_df)
-        #     except Exception as e:
-        #         logger.exception(f"{e}, skipping to next day.")
-        #         continue
 
 
 if __name__ == '__main__':
